Cleanup and rewrite of usb modes, to change configurations easily #3
Added support for alternative Network mode (rndis) which will be
used for USB tethering.

Removed weird mode wrappings, available modes are only using
ro.properties and not specific devices version. This job is only
made in Activity to hide unavailable choices to support them if
user or another tool has set the device is this mode.

Should now be easier to tune for different devices

Signed-off-by: Tanguy Pruvot <tanguy.pruvot@gmail.com>

Change-Id: I3753ac61c806fdb0be9b2305d83a8f0a332ddce5