Update patch set 6
Patch Set 6:

Actually, to use getApplicationContext() instead of UsbService.this was the first thing I tried to get it fixed, but it made no change. And I think that UsbService.this is more specific context, therefore the preferred one.

The only case I observe the toast not displayed is when switching from Memory Card Access to Charge Only. When changing from Portal & Tools to Charge Only, it is OK, as well as when the usb cable is connected and Charge Only mode activated.

The toast seems to be correctly enqueued, but not shown in this particular case.

Patch-set: 6