Update patch set 8
Patch Set 8:

In this patch set, I've delayed the toast until after the enumeration took place. The root cause is that IMountService.setUsbMassStorageEnabled(false) causes the SD card to be mounted again, which makes the call take a long time, thus blocking the main loop of the service, thus causing the toast to be hidden at the same time it's shown.

Previously, Moto seemed to send a 'show the toast' broadcast to itself, which seems to avoid the problem ... I have no explanation for that, though, and want to avoid the additional overhead caused by sending a broadcast, at least until somebody can explain to me _why_ it works with a broadcast. If there's no good explanation, it might just have worked by accident.

Downside of this is that there now is a delay between selecting the mode and the toast, but that maybe just is a matter of getting used to.

Patch-set: 8